refactor hierarchy tree walker into hierarchy panel
