+++ packages/editor/src/panels/hierarchy/helpers.ts
@@ -23,14 +23,31 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
+import { GLTF } from '@gltf-transform/core'
 import { NotificationService } from '@ir-engine/client-core/src/common/services/NotificationService'
-import { Entity, getComponent, UUIDComponent } from '@ir-engine/ecs'
+import { Entity, entityExists, EntityUUID, getComponent, hasComponent, UUIDComponent } from '@ir-engine/ecs'
 import { AllFileTypes } from '@ir-engine/engine/src/assets/constants/fileTypes'
+import { GLTFSnapshotState } from '@ir-engine/engine/src/gltf/GLTFState'
+import { ModelComponent } from '@ir-engine/engine/src/scene/components/ModelComponent'
+import { SourceComponent } from '@ir-engine/engine/src/scene/components/SourceComponent'
+import { getModelSceneID } from '@ir-engine/engine/src/scene/functions/loaders/ModelFunctions'
 import { getState } from '@ir-engine/hyperflux'
 import { t } from 'i18next'
 import { CopyPasteFunctions } from '../../functions/CopyPasteFunctions'
 import { EditorControlFunctions } from '../../functions/EditorControlFunctions'
+import { HierarchyTreeState } from '../../services/HierarchyNodeState'
 import { SelectionState } from '../../services/SelectionServices'
+
+export type HierarchyTreeNodeType = {
+  depth: number
+  entity: Entity
+  childIndex: number
+  lastChild: boolean
+  isLeaf?: boolean
+  isCollapsed?: boolean
+}
+
+type NestedHierarchyTreeNode = HierarchyTreeNodeType & { children: NestedHierarchyTreeNode[] }
 
 /* COMMON */
 
@@ -40,6 +57,8 @@ export const uploadOptions = {
 }
 
 /** UTILITIES **/
+
+/* NODE FUNCTIONALITIES */
 
 const getSelectedEntities = (entity?: Entity) => {
   const selected = entity
@@ -77,4 +96,125 @@ export const pasteNodes = (entity?: Entity) => {
         variant: 'error'
       })
     })
+}
+
+/* HIERARCHY TREE WALKER */
+
+function isChild(index: number, nodes: GLTF.INode[]) {
+  for (const node of nodes) {
+    if (node.children && node.children.includes(index)) return true
+  }
+
+  return false
+}
+
+function buildHierarchyTree(
+  depth: number,
+  childIndex: number,
+  node: GLTF.INode,
+  nodes: GLTF.INode[],
+  array: NestedHierarchyTreeNode[],
+  lastChild: boolean,
+  sceneID: string,
+  showModelChildren: boolean
+) {
+  const uuid = node.extensions && (node.extensions[UUIDComponent.jsonID] as EntityUUID)
+  const entity = UUIDComponent.getEntityByUUID(uuid!)
+  if (!entity || !entityExists(entity)) return
+
+  const item = {
+    depth,
+    childIndex,
+    entity: entity,
+    isCollapsed: !getState(HierarchyTreeState).expandedNodes[sceneID]?.[entity],
+    children: [],
+    isLeaf: !(node.children && node.children.length > 0),
+    lastChild: lastChild
+  }
+  array.push(item)
+
+  if (hasComponent(entity, ModelComponent) && showModelChildren) {
+    const modelSceneID = getModelSceneID(entity)
+    const snapshotState = getState(GLTFSnapshotState)
+    const snapshots = snapshotState[modelSceneID]
+    if (snapshots) {
+      const snapshotNodes = snapshots.snapshots[snapshots.index].nodes
+      if (snapshotNodes && snapshotNodes.length > 0) {
+        item.isLeaf = false
+        if (!item.isCollapsed)
+          buildHierarchyTreeForNodes(depth + 1, snapshotNodes, item.children, sceneID, showModelChildren)
+      }
+    }
+  }
+
+  if (node.children && !item.isCollapsed) {
+    for (let i = 0; i < node.children.length; i++) {
+      const childIndex = node.children[i]
+      buildHierarchyTree(
+        depth + 1,
+        i,
+        nodes[childIndex],
+        nodes,
+        item.children,
+        i === node.children.length - 1,
+        sceneID,
+        showModelChildren
+      )
+    }
+  }
+}
+
+function buildHierarchyTreeForNodes(
+  depth: number,
+  nodes: GLTF.INode[],
+  outArray: NestedHierarchyTreeNode[],
+  sceneID: string,
+  showModelChildren: boolean
+) {
+  for (let i = 0; i < nodes.length; i++) {
+    if (isChild(i, nodes)) continue
+    buildHierarchyTree(depth, i, nodes[i], nodes, outArray, false, sceneID, showModelChildren)
+  }
+  if (!outArray.length) return
+  outArray[outArray.length - 1].lastChild = true
+}
+
+function flattenTree(array: NestedHierarchyTreeNode[], outArray: HierarchyTreeNodeType[]) {
+  for (const item of array) {
+    if (!item.entity) continue
+    outArray.push({
+      depth: item.depth,
+      entity: item.entity,
+      childIndex: item.childIndex,
+      lastChild: item.lastChild,
+      isLeaf: item.isLeaf,
+      isCollapsed: item.isCollapsed
+    })
+    flattenTree(item.children, outArray)
+  }
+}
+
+export function gltfHierarchyTreeWalker(
+  rootEntity: Entity,
+  nodes: GLTF.INode[],
+  showModelChildren: boolean
+): HierarchyTreeNodeType[] {
+  const outArray = [] as NestedHierarchyTreeNode[]
+
+  const sceneID = getComponent(rootEntity, SourceComponent)
+  const rootNode = {
+    depth: 0,
+    entity: rootEntity,
+    childIndex: 0,
+    lastChild: true,
+    isCollapsed: !getState(HierarchyTreeState).expandedNodes[sceneID]?.[rootEntity]
+  }
+  const tree = [rootNode] as HierarchyTreeNodeType[]
+
+  if (!rootNode.isCollapsed) {
+    buildHierarchyTreeForNodes(1, nodes, outArray, sceneID, showModelChildren)
+    flattenTree(outArray, tree)
+  }
+
+  return tree
 }
